Sorting array (Insertin Sort Algorithm)
A  menu driven program as given below, to sort an array of n integers in ascending
order by insertion sort algorithm and time required (in terms of
step/frequency count) to sort the elements. Repeat the experiment for different values of
n and different nature of data ( i.e. apply insertion sort algorithm on the data of array
that are already sorted, reversely sorted and random data). Finally plot a graph of the
time taken versus n for each type of data. The elements can be read from a file or can be
generated using the random number generator.